fix(memory): keep migration safeguards portable
