+++ crates/velesdb-memory/Cargo.toml
@@ -96,9 +96,10 @@ hyper-util = { version = "0.1", optional = true, features = ["tokio", "server-au
 # for a single request inside the manual TLS accept loop.
 tower = { workspace = true, optional = true }
 
-[target.'cfg(windows)'.dependencies.windows-sys]
-version = "0.61"
-features = ["Win32_Foundation", "Win32_Storage_FileSystem"]
+[target.'cfg(windows)'.dependencies.atomicwrites]
+# Safe wrapper around MoveFileExW(REPLACE_EXISTING | WRITE_THROUGH), keeping
+# the crate's own code free of unsafe while retaining a durable replacement.
+version = "0.4.4"
 optional = true
 
 [features]
@@ -143,7 +144,7 @@ persistence = [
     "velesdb-core/persistence",
     "dep:sha2",
     "dep:fs2",
-    "dep:windows-sys",
+    "dep:atomicwrites",
 ]
 # Real on-device semantic recall via a local Ollama embeddings endpoint. Off by
 # default so the shipped binary stays tiny, zero-dependency, and fully offline.
